Support 96 bit nonces, and a `c.Seek(counter)` method.
This is the IETF RFC 7539 style ChaCha20 with a 96 bit nonce and 32 bit
counter, used for constructing `AEAD_CHACHA20_POLY1305` and HS1-SIV
amongst other things.

The code will panic when the 32 bit counter wraps, because that's the
easy thing to do, and 2^32 blocks per key/nonce should be sufficient
for anybody, though I reserve the right to change this behavior as
needed at a later date.  Poly1305 only allows authenticating
2^64 bytes anyway so in practice this should never be a concern.